+++ src/test/java/org/realityforge/bazel/depgen/MainTest.java
@@ -39,10 +39,10 @@ public class MainTest
     assertEquals( handler.toString(),
                   "java org.realityforge.bazel.depgen.Main [options] [command]\n" +
                   "\tPossible Commands:\n" +
-                  "\t\tgenerate: Generate the bazel extension from the dependency configuration.\n" +
+                  "\t\tgenerate: Generate the Bazel outputs from the dependency configuration.\n" +
                   "\t\tprint-graph: Compute and print the dependency graph for the dependency configuration.\n" +
                   "\t\thash: Generate a hash of the content of the dependency configuration.\n" +
-                  "\t\tinit: Initialize an empty dependency configuration and workspace infrastructure.\n" +
+                  "\t\tinit: Initialize an empty dependency configuration and Bazel scaffolding.\n" +
                   "\t\tinfo: Print runtime info about the tool.\n" +
                   "\tOptions:\n" +
                   "\t--version\n" +