+++ src/test/java/org/realityforge/bazel/depgen/model/OptionsModelTest.java
@@ -7,6 +7,8 @@ import org.realityforge.bazel.depgen.config.NameStrategy;
 import org.realityforge.bazel.depgen.config.GlobalJavaConfig;
 import org.realityforge.bazel.depgen.config.Nature;
 import org.realityforge.bazel.depgen.config.OptionsConfig;
+import org.realityforge.bazel.depgen.config.RepositoryRuleGenerationStrategy;
+import org.realityforge.bazel.depgen.config.TargetGenerationStrategy;
 import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
@@ -26,6 +28,13 @@ public class OptionsModelTest
                   FileUtil.getCurrentDirectory().resolve( OptionsConfig.DEFAULT_EXTENSION_FILE ) );
     assertEquals( model.getWorkspaceMacroName(), OptionsConfig.DEFAULT_WORKSPACE_MACRO_NAME );
     assertEquals( model.getTargetMacroName(), OptionsConfig.DEFAULT_TARGET_MACRO_NAME );
+    assertEquals( model.getRepositoryRuleGenerationStrategy(),
+                  OptionsConfig.DEFAULT_REPOSITORY_RULE_GENERATION_STRATEGY );
+    assertEquals( model.getTargetGenerationStrategy(), OptionsConfig.DEFAULT_TARGET_GENERATION_STRATEGY );
+    assertEquals( model.getRepositoryRuleStartToken(), OptionsConfig.DEFAULT_REPOSITORY_RULE_START_TOKEN );
+    assertEquals( model.getRepositoryRuleEndToken(), OptionsConfig.DEFAULT_REPOSITORY_RULE_END_TOKEN );
+    assertEquals( model.getTargetStartToken(), OptionsConfig.DEFAULT_TARGET_START_TOKEN );
+    assertEquals( model.getTargetEndToken(), OptionsConfig.DEFAULT_TARGET_END_TOKEN );
     assertEquals( model.getNamePrefix(), OptionsConfig.DEFAULT_NAME_PREFIX );
     assertEquals( model.getNameStrategy(), OptionsConfig.DEFAULT_NAME_STRATEGY );
     assertEquals( model.getRepositoryNameStrategy(), OptionsConfig.DEFAULT_REPOSITORY_NAME_STRATEGY );
@@ -77,6 +86,12 @@ public class OptionsModelTest
     source.setExtensionFile( "dependencies.bzl" );
     source.setWorkspaceMacroName( "gen_myprj_dependency_rules" );
     source.setTargetMacroName( "gen_myprj_targets" );
+    source.setRepositoryRuleGenerationStrategy( "module" );
+    source.setTargetGenerationStrategy( "build" );
+    source.setRepositoryRuleStartToken( "# rs" );
+    source.setRepositoryRuleEndToken( "# re" );
+    source.setTargetStartToken( "# ts" );
+    source.setTargetEndToken( "# te" );
     source.setNamePrefix( "myprj_" );
     source.setNameStrategy( NameStrategy.ArtifactId );
     source.setRepositoryNameStrategy( NameStrategy.ArtifactId );
@@ -98,6 +113,12 @@ public class OptionsModelTest
     assertEquals( model.getExtensionFile(), thirdpartyDir.resolve( "dependencies.bzl" ) );
     assertEquals( model.getWorkspaceMacroName(), "gen_myprj_dependency_rules" );
     assertEquals( model.getTargetMacroName(), "gen_myprj_targets" );
+    assertEquals( model.getRepositoryRuleGenerationStrategy(), RepositoryRuleGenerationStrategy.Module );
+    assertEquals( model.getTargetGenerationStrategy(), TargetGenerationStrategy.Build );
+    assertEquals( model.getRepositoryRuleStartToken(), "# rs" );
+    assertEquals( model.getRepositoryRuleEndToken(), "# re" );
+    assertEquals( model.getTargetStartToken(), "# ts" );
+    assertEquals( model.getTargetEndToken(), "# te" );
     assertEquals( model.getNamePrefix(), "myprj_" );
     assertEquals( model.getNameStrategy(), NameStrategy.ArtifactId );
     assertEquals( model.getRepositoryNameStrategy(), NameStrategy.ArtifactId );
@@ -109,5 +130,33 @@ public class OptionsModelTest
     assertTrue( model.exportDeps() );
     assertTrue( model.supportDependencyOmit() );
     assertFalse( model.verifyConfigSha256() );
+  }
+
+  @Test
+  public void parseWithInvalidRepositoryRuleGenerationStrategy()
+  {
+    final OptionsConfig source = new OptionsConfig();
+    source.setRepositoryRuleGenerationStrategy( "bogus" );
+
+    final InvalidModelException exception =
+      expectThrows( InvalidModelException.class,
+                    () -> OptionsModel.parse( FileUtil.getCurrentDirectory(), source ) );
+    assertEquals( exception.getMessage(),
+                  "The options.repositoryRuleGenerationStrategy property must be one of: extensionFile, module." );
+    assertEquals( exception.getModel(), source );
+  }
+
+  @Test
+  public void parseWithInvalidTargetGenerationStrategy()
+  {
+    final OptionsConfig source = new OptionsConfig();
+    source.setTargetGenerationStrategy( "bogus" );
+
+    final InvalidModelException exception =
+      expectThrows( InvalidModelException.class,
+                    () -> OptionsModel.parse( FileUtil.getCurrentDirectory(), source ) );
+    assertEquals( exception.getMessage(),
+                  "The options.targetGenerationStrategy property must be one of: extensionFile, build." );
+    assertEquals( exception.getModel(), source );
   }
 }